LayoutTests/imported/mozilla/css-transitions/test_keyframeeffect-getkeyframes-expected.txt


FAIL KeyframeEffectReadOnly.getKeyframes() returns expected frames for a simple transition assert_equals: value for 'easing' on ComputedKeyframe #0 expected "ease" but got "linear"
FAIL KeyframeEffectReadOnly.getKeyframes() returns expected frames for a simple transition with a non-default easing function assert_equals: value for 'easing' on ComputedKeyframe #0 expected "steps(2)" but got "linear"
FAIL KeyframeEffectReadOnly.getKeyframes() returns expected frames for a transition with a CSS variable endpoint assert_equals: value for 'easing' on ComputedKeyframe #0 expected "ease" but got "linear"
